% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test}
\usage{
test(test = TRUE)
}
\arguments{
\item{test}{test.}
}
\value{
test
}
\description{
test
}
\details{
\insertCite{Reyna2022}{EpiVec}
}
\examples{
print test

}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr})
}
